requirements.txt
pandas
geopandas
plotly
nbformat==4.2.0
boto3
python-dotenv
dash
dash-bootstrap-components
awscli